댓글 주인만 삭제 가능

+++ src/main/java/com/caprj240521backend/controller/comment/CommentController.java
@@ -3,6 +3,7 @@ package com.caprj240521backend.controller.comment;
 import com.caprj240521backend.domain.comment.Comment;
 import com.caprj240521backend.service.comment.CommentService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
@@ -34,7 +35,12 @@ public class CommentController {
     }
 
     @DeleteMapping("remove")
-    public void remove(@RequestBody Comment comment) {
-        service.remove(comment);
+    public ResponseEntity remove(@RequestBody Comment comment, Authentication authentication) {
+        if (service.hasAccess(comment, authentication)) {
+            service.remove(comment);
+            return ResponseEntity.ok().build();
+        } else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
     }
 }
